+++ src/main/java/com/moratan251/psitweaks/client/gui/machine/GuiTranscendentEnergyCube.java
@@ -1,0 +1,49 @@
+package com.moratan251.psitweaks.client.gui.machine;
+
+import com.moratan251.psitweaks.common.tile.machine.TileEntityTranscendentEnergyCube;
+import mekanism.client.SpecialColors;
+import mekanism.client.gui.GuiConfigurableTile;
+import mekanism.client.gui.element.GuiSideHolder;
+import mekanism.client.gui.element.gauge.GaugeType;
+import mekanism.client.gui.element.gauge.GuiEnergyGauge;
+import mekanism.client.gui.element.tab.GuiEnergyTab;
+import mekanism.client.gui.element.tab.GuiSecurityTab;
+import mekanism.common.MekanismLang;
+import mekanism.common.inventory.container.tile.MekanismTileContainer;
+import mekanism.common.util.text.EnergyDisplay;
+import net.minecraft.client.gui.GuiGraphics;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.player.Inventory;
+
+import java.util.List;
+
+public class GuiTranscendentEnergyCube extends GuiConfigurableTile<TileEntityTranscendentEnergyCube, MekanismTileContainer<TileEntityTranscendentEnergyCube>> {
+
+    public GuiTranscendentEnergyCube(MekanismTileContainer<TileEntityTranscendentEnergyCube> container, Inventory inv, Component title) {
+        super(container, inv, title);
+        dynamicSlots = true;
+    }
+
+    @Override
+    protected void addGuiElements() {
+        addRenderableWidget(GuiSideHolder.create(this, imageWidth, 36, 98, false, true, SpecialColors.TAB_ARMOR_SLOTS));
+        super.addGuiElements();
+        addRenderableWidget(new GuiEnergyGauge(tile.getEnergyContainer(), GaugeType.WIDE, this, 55, 18));
+        addRenderableWidget(new GuiEnergyTab(this, () -> List.of(
+                MekanismLang.MATRIX_INPUT_RATE.translate(EnergyDisplay.of(tile.getInputRate())),
+                MekanismLang.MAX_OUTPUT.translate(EnergyDisplay.of(TileEntityTranscendentEnergyCube.getOutput()))
+        )));
+    }
+
+    @Override
+    protected void addSecurityTab() {
+        addRenderableWidget(new GuiSecurityTab(this, tile, 6));
+    }
+
+    @Override
+    protected void drawForegroundText(GuiGraphics guiGraphics, int mouseX, int mouseY) {
+        renderTitleText(guiGraphics);
+        drawString(guiGraphics, playerInventoryTitle, inventoryLabelX, inventoryLabelY, titleTextColor());
+        super.drawForegroundText(guiGraphics, mouseX, mouseY);
+    }
+}