+++ src/main/java/com/moratan251/psitweaks/common/registries/PsitweaksMekanismBlocks.java
@@ -2,10 +2,12 @@ package com.moratan251.psitweaks.common.registries;
 
 import com.moratan251.psitweaks.Psitweaks;
 import com.moratan251.psitweaks.common.blocks.transmitter.BlockTranscendentCable;
+import com.moratan251.psitweaks.common.items.block.ItemBlockTranscendentEnergyCube;
 import com.moratan251.psitweaks.common.items.block.ItemBlockTranscendentCable;
 import com.moratan251.psitweaks.common.tile.machine.TileEntityMaterialMutator;
 import com.moratan251.psitweaks.common.tile.machine.TileEntityPsionicGenerator;
 import com.moratan251.psitweaks.common.tile.machine.TileEntitySculkEroder;
+import com.moratan251.psitweaks.common.tile.machine.TileEntityTranscendentEnergyCube;
 import mekanism.common.block.prefab.BlockTile;
 import mekanism.common.content.blocktype.BlockTypeTile;
 import mekanism.common.item.block.machine.ItemBlockMachine;
@@ -34,6 +36,11 @@ public class PsitweaksMekanismBlocks {
 
     public static final BlockRegistryObject<BlockTranscendentCable, ItemBlockTranscendentCable> TRANSCENDENT_CABLE =
             BLOCKS.register("transcendent_universal_cable", BlockTranscendentCable::new, ItemBlockTranscendentCable::new);
+
+    public static final BlockRegistryObject<BlockTile.BlockTileModel<TileEntityTranscendentEnergyCube, BlockTypeTile<TileEntityTranscendentEnergyCube>>, ItemBlockTranscendentEnergyCube> TRANSCENDENT_ENERGY_CUBE =
+            BLOCKS.register("transcendent_energy_cube",
+                    () -> new BlockTile.BlockTileModel<>(PsitweaksMekanismBlockTypes.TRANSCENDENT_ENERGY_CUBE, properties -> properties),
+                    ItemBlockTranscendentEnergyCube::new);
 
     private PsitweaksMekanismBlocks() {
     }